services/pricing-service/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
setuptools>=75.0.0,<81
asyncpg==0.30.0
azure-identity==1.19.0
aiohttp==3.10.10
pydantic==2.9.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-asyncpg==0.48b0
python-json-logger==2.0.7
